server/ai_service/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
imagehash
torch
transformers
